Add CORS header to allow serving of files from different hosts
I'm having an application running on port 8500 and serve files which are built using `webpack-serve` in CLI mode on port 8080. That results in an error:
> _Failed to load http://localhost:8080/static/70a4aeb6b1596bc852d3.hot-update.json: No 'Access-Control-Allow-Origin' header is present on the requested resource. Origin 'http://localhost:8500' is therefore not allowed access._

I was wondering if I'm using it incorrectly since I found nobody with similar issues here in the issue board but adding a simple `Access-Control-Allow-Origin: *` header fixed it for me. Since `webpack-serve` is typically only used in dev environments I don't think this is critical.